Move code from totalAnalysis to totalFuncs
